refactor misc section to AccountMiscBox component
