+++ client/src/pages/Account/index.js
@@ -1,3 +1,4 @@
 export { default as AccountSettings } from './AccountSettings';
 export { default as AccountDetailsBox } from './AccountDetailsBox';
 export { default as PasswordUpdateBox } from './PasswordUpdateBox';
+export { default as AccountMiscBox } from './AccountMiscBox';
